@@ -94,51 +92,38 @@
     "rimraf": "5.0.5",
     "safe-compare": "1.1.4",
     "sanitize-filename": "1.6.3",
-    "sanitize-html": "2.11.0",
+    "sanitize-html": "2.12.1",
     "sax": "1.3.0",
-    "semver": "7.5.4",
+    "semver": "7.6.0",
     "serve-favicon": "2.5.0",
     "session-file-store": "1.5.0",
     "split.js": "1.6.5",
     "stream-throttle": "0.1.3",
     "striptags": "3.2.0",
-    "tmp": "0.2.1",
+    "tmp": "0.2.3",
     "tree-kill": "1.2.2",
     "turndown": "7.1.2",
     "unescape": "1.0.1",
-    "ws": "8.14.2",
+    "ws": "8.16.0",
     "xml2js": "0.6.2",
-    "yauzl": "2.10.0"
+    "yauzl": "3.1.2"
   },
   "devDependencies": {
     "cross-env": "7.0.3",
     "electron": "25.9.8",
-    "electron-builder": "24.6.4",
+    "electron-builder": "24.13.3",
     "electron-packager": "17.1.2",
     "electron-rebuild": "3.2.9",
-    "eslint": "8.54.0",
-    "eslint-config-airbnb-base": "15.0.0",
-    "eslint-config-prettier": "9.0.0",
-    "eslint-plugin-import": "2.29.0",
-    "eslint-plugin-jsonc": "2.10.0",
-    "eslint-plugin-prettier": "5.0.1",
     "esm": "3.2.25",
-    "husky": "8.0.3",
     "jasmine": "5.1.0",
     "jsdoc": "4.0.2",
-    "jsonc-eslint-parser": "2.4.0",
-    "lint-staged": "15.1.0",
     "lorem-ipsum": "2.0.8",
-    "nodemon": "3.0.1",
-    "prettier": "3.1.0",
+    "nodemon": "3.1.0",
     "rcedit": "4.0.1",
-    "webpack": "5.89.0",
+    "webpack": "5.90.3",
     "webpack-cli": "5.1.4"
   },
   "optionalDependencies": {
     "electron-installer-debian": "3.2.0"
-  },
-  "lint-staged": {
-    "*.js": "eslint --cache --fix"
   }
 }
