@@ -105,10 +105,10 @@ function renderText(result, note) {
 
         if (result.content.includes(`<span class="math-tex">`)) {
             result.header += `
-<script src="../../${assetPath}/node_modules/katex/dist/katex.min.js"></script>
-<link rel="stylesheet" href="../../${assetPath}/node_modules/katex/dist/katex.min.css">
-<script src="../../${assetPath}/node_modules/katex/dist/contrib/auto-render.min.js"></script>
-<script src="../../${assetPath}/node_modules/katex/dist/contrib/mhchem.min.js"></script>
+<script src="../${assetPath}/node_modules/katex/dist/katex.min.js"></script>
+<link rel="stylesheet" href="../${assetPath}/node_modules/katex/dist/katex.min.css">
+<script src="../${assetPath}/node_modules/katex/dist/contrib/auto-render.min.js"></script>
+<script src="../${assetPath}/node_modules/katex/dist/contrib/mhchem.min.js"></script>
 <script>
 document.addEventListener("DOMContentLoaded", function() {
     renderMathInElement(document.getElementById('content'));
